fix forwarder creation bug on ams 1.7.1, fix cmake project version bug
i was checking against ams 1.7.1, however the change was introduced in ams 1.8.0.
fixes #106

as for cmake, i have no idea why this bug happens, but sometimes it will use the project version for hbl.
which results in sphaira having a version number of 3.0.0.
this only seemed to happen when building the zip with no current build.